+++ src/components/Header.js
@@ -4,7 +4,6 @@ import GoogleAuth from './GoogleAuth';
 const Header = ({ isSignedIn, updateSignIn, updateUserId }) => {
   return (
     <div className="ui secondary pointing menu">
-      {/* switch to Link */}
       <div className="ui header">PostIt</div>
       <div className="right menu">
         <GoogleAuth
