+++ src/components/PostDetails.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Link} from 'react-router-dom';
-import './post-details.css';
+import '../css/post-details.css';
 
 const PostDetails = (props) => {
    const posts = props.posts;
@@ -14,8 +14,6 @@ const PostDetails = (props) => {
       <div className="post-main-div container">
         <div className="post-top-div content">
           <Link to='/'>Back to Posts</Link>
-          {/*<button className="postButton">Delete Post</button>*/}
-          {/*<button className="postButton1">Edit Post</button>*/}
         </div>
         <div className="post-text-div header">
           <h4>Title: {filteredPost.title}</h4>
