+++ src/css/post-details.css
@@ -7,7 +7,7 @@
   display: flex;
   flex-direction: column;
   width: 50%;
-  margin: auto;
+  margin: 20px auto;
   border: 2px solid black;
 }
 .post-top-div {
